feat: add image uploading, deleting and updating functionality to relevant routes

+++ src/server/api/routers/zenith-media.ts
@@ -2,6 +2,7 @@ import { revalidateTag } from "next/cache";
 import { z } from "zod";
 import { deleteFileFromSftpServer } from "~/app/api/sftp/utils/sftp-engine";
 import { MIN_MEDIA_ORDER_NUMBER } from "~/constants/zenith-media";
+import { env } from "~/env.mjs";
 import { objectId } from "~/schemas/helpers/custom-zod-helpers";
 import {
   createZenithMediaServerSchema,
@@ -98,14 +99,16 @@ export const zenithMediaRouter = createTRPCRouter({
         throw new Error("Media not found");
       }
 
-      try {
-        await deleteFileFromSftpServer({ url: fileUrl.url });
-      } catch (error) {
-        if (error instanceof Error) {
-          if (error.message !== "Filen du ville radera kunde inte hittas!")
-            throw new Error(error.message);
-        } else {
-          throw new Error("Something went wrong when deleting the file.");
+      if (fileUrl.url.startsWith(env.NEXT_PUBLIC_SFTP_BASE_URL)) {
+        try {
+          await deleteFileFromSftpServer({ url: fileUrl.url });
+        } catch (error) {
+          if (error instanceof Error) {
+            if (error.message !== "Filen du ville radera kunde inte hittas!")
+              throw new Error(error.message);
+          } else {
+            throw new Error("Something went wrong when deleting the file.");
+          }
         }
       }
 
@@ -114,6 +117,14 @@ export const zenithMediaRouter = createTRPCRouter({
           id: id,
         },
       });
+
+      // Delete image if it exists
+      if (
+        deletedMedia.coverImage &&
+        deletedMedia.coverImage.startsWith(env.NEXT_PUBLIC_SFTP_BASE_URL)
+      ) {
+        await deleteFileFromSftpServer({ url: deletedMedia.coverImage });
+      }
 
       revalidateTag("zenithMedia");
